Thu, Jun 29, 2023, 11:03 PM +02:00

+++ week04/pset4/filter-less/helpers.c
@@ -73,9 +73,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     // Calculate blur value of pixel
     for (int i = 0; i < height; i++)
     {
-        long blurredRed = 0;
-        long blurredGreen = 0;
-        long blurredBlue = 0;
+        int blurredRed = 0;
+        int blurredGreen = 0;
+        int blurredBlue = 0;
         int divisor = 0;
 
         for (int j = 0; j < width; j++)
@@ -91,9 +91,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
             }
             // Save average color to image
-            image[i][j].rgbtRed = (BYTE) round(blurredRed / divisor);
-            image[i][j].rgbtGreen = (BYTE) round(blurredGreen  / divisor);
-            image[i][j].rgbtBlue = (BYTE) round(blurredBlue / divisor);
+            image[i][j].rgbtRed = round(blurredRed / divisor);
+            image[i][j].rgbtGreen = round(blurredGreen  / divisor);
+            image[i][j].rgbtBlue = round(blurredBlue / divisor);
 
             // Reset colors
             blurredRed = 0;
